fix(analytics): cast double precision to numeric before ROUND() in category-velocity
PostgreSQL only has round(numeric, integer), not round(double precision, integer).

+++ backend/app/routers/analytics.py
@@ -1878,12 +1878,12 @@ async def category_velocity(
     velocity_row = (await db.execute(sql_text("""
         SELECT
             COUNT(*) AS total_sold,
-            ROUND(AVG(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0), 1) AS avg_days,
-            ROUND(MIN(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0), 1) AS min_days,
-            ROUND(MAX(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0), 1) AS max_days,
-            ROUND(AVG(l.price), 0) AS avg_price,
-            ROUND(PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY l.price), 0) AS p25_price,
-            ROUND(PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY l.price), 0) AS p75_price
+            ROUND(AVG(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0)::numeric, 1) AS avg_days,
+            ROUND(MIN(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0)::numeric, 1) AS min_days,
+            ROUND(MAX(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0)::numeric, 1) AS max_days,
+            ROUND(AVG(l.price)::numeric, 0) AS avg_price,
+            ROUND(PERCENTILE_CONT(0.25) WITHIN GROUP (ORDER BY l.price)::numeric, 0) AS p25_price,
+            ROUND(PERCENTILE_CONT(0.75) WITHIN GROUP (ORDER BY l.price)::numeric, 0) AS p75_price
         FROM auctions a
         INNER JOIN listings l ON l.id = a.listing_id
         WHERE a.status = 'ended'
@@ -1897,7 +1897,7 @@ async def category_velocity(
     price_sens = (await db.execute(sql_text("""
         SELECT
             CASE WHEN l.price <= pct.p50 THEN 'ucuz' ELSE 'pahalı' END AS bucket,
-            ROUND(AVG(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0), 1) AS avg_days,
+            ROUND(AVG(EXTRACT(EPOCH FROM (COALESCE(a.ended_at, a.started_at) - l.created_at)) / 86400.0)::numeric, 1) AS avg_days,
             COUNT(*) AS count
         FROM auctions a
         INNER JOIN listings l ON l.id = a.listing_id
